requirements.txt
Django
djangorestframework
mysqlclient
pytz
sqlparse
django-cors-headers
django-utils-six
dj-static
drf-yasg